# Conflicts:
#	.gradle/7.4/executionHistory/executionHistory.bin
#	.gradle/7.4/executionHistory/executionHistory.lock
#	.gradle/7.4/fileHashes/fileHashes.bin
#	.gradle/7.4/fileHashes/fileHashes.lock
#	.gradle/7.4/fileHashes/resourceHashesCache.bin
#	.gradle/buildOutputCleanup/buildOutputCleanup.lock
#	.gradle/file-system.probe
#	build/classes/java/main/Main3.class
#	build/classes/java/main/entities/Item.class
#	build/classes/java/main/screens/MainEmployeeScreen.class
#	build/classes/java/main/use_cases/arr/Add.class
#	build/classes/java/main/use_cases/arr/RemoveUseCase.class
#	build/classes/java/main/use_cases/arr/ReturnUseCase.class
#	build/tmp/compileJava/previous-compilation-data.bin
#	src/main/java/use_cases/arr/Add.java
#	src/main/java/use_cases/arr/RemoveUseCase.java
#	src/main/java/use_cases/arr/ReturnUseCase.java

+++ src/main/java/use_cases/sales_report/SalesReporterInputData.java
@@ -102,7 +102,7 @@ public class SalesReporterInputData {
         int quantity = 0;
         for (String[] row: rows){
             if (row[ACTION_COLUMN].equals("SELL ITEM") && row[SERIAL_NUM_COLUMN].equals(serialNum)){
-                quantity += Integer.parseInt(row[QUANTITY_COLUMN]);
+                quantity += Double.parseDouble(row[QUANTITY_COLUMN]);
             }
         }
         return quantity;
